Travis: don't use tox, don't run tests in parallel
Sometimes, when the sandbox is under heavy load, we get timeout errors.
Be easy on the sandbox by not running the tests in parallel. The speed
we get by running the tests in parallel does not compensate the
troubling random errors.

Also, don't need to use tox in travis.

Also, run flake8 only over the project files.